feat(utils): add date helper
