+++ src/sele_saisie_auto/saisie_automatiser_psatime.py
@@ -61,11 +61,9 @@ from sele_saisie_auto.selenium_utils import set_log_file as set_log_file_seleniu
 from sele_saisie_auto.selenium_utils import wait_for_dom_after
 from sele_saisie_auto.shared_memory_service import SharedMemoryService
 from sele_saisie_auto.timeouts import DEFAULT_TIMEOUT
+from sele_saisie_auto.utils.date_utils import get_next_saturday_if_not_saturday
 from sele_saisie_auto.utils.misc import program_break_time
-from sele_saisie_auto.utils.mission import (
-    est_en_mission,
-    get_next_saturday_if_not_saturday,
-)
+from sele_saisie_auto.utils.mission import est_en_mission
 
 # ----------------------------------------------------------------------------- #
 # ------------------------------- CONSTANTE ----------------------------------- #
