+++ src/sele_saisie_auto/utils/__init__.py
@@ -1,3 +1,3 @@
 """Utility subpackages."""
 
-__all__ = ["misc", "mission"]
+__all__ = ["misc", "mission", "date_utils"]
